voicemail-agent/requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
dataclasses-json==0.6.7
deepgram-sdk==3.11.0
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastui==0.7.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jiter==0.9.0
marshmallow==3.26.1
multidict==6.4.3
mypy-extensions==1.0.0
openai==1.74.0
packaging==24.2
propcache==0.3.1
pydantic==2.9.2
pydantic_core==2.33.1
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
websockets==15.0.1
yarl==1.19.0
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
dataclasses-json==0.6.7
deepgram-sdk==3.11.0
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastui==0.7.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jiter==0.9.0
marshmallow==3.26.1
multidict==6.4.3
mypy-extensions==1.0.0
openai==1.74.0
packaging==24.2
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
websockets==15.0.1
yarl==1.19.0
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
click==8.1.8
dataclasses-json==0.6.7
deepgram-sdk==3.11.0
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastui==0.7.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jiter==0.9.0
marshmallow==3.26.1
multidict==6.4.3
mypy-extensions==1.0.0
openai==1.74.0
packaging==24.2
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.1
websockets==15.0.1
yarl==1.19.0
